refactor(statement): ajustes em Statement service e controller
- StatementService: melhorias na busca e criação
- SubjectService: ajustes para integração com OCR
- StatementController: endpoints atualizados
- StatementBasicResponse: campos adicionais

+++ services/backend-api/src/main/java/ao/creativemode/kixi/dto/statement/StatementBasicResponse.java
@@ -1,11 +1,10 @@
 package ao.creativemode.kixi.dto.statement;
 
 public record StatementBasicResponse(
-        Long id,
-        String examType,
-        String variant,
-        String title,
-        Integer durationMinutes,
-        Integer totalMaxScore
-) {
-}
+    Long id,
+    String examType,
+    String variant,
+    String title,
+    Integer durationMinutes,
+    Double totalMaxScore
+) {}
